requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.2
gensim==4.3.3
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.7.0
scipy==1.13.1
six==1.17.0
smart-open==7.1.0
soupsieve==2.7
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wrapt==1.17.2
